requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.31
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.2.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sounddevice==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wrapt==2.0.1
